Rename core modules to match class names
