+++ tests/unit/core/DependencyAwareClass.spec.ts
