+++ tests/unit/core/DependencyInjection.spec.ts
